packetdrill: test mtu probe with tx timestamp
When building a large mtu probe packet, TCP may coalesce packets.
Verify that it also correctly coalesces the tx timestamp request.

This exposes a bug due to missing support in tcp_mtu_probe. Though
with the following change to that function:

    -       size_needed = probe_size + (tp->reordering + 1) * tp->mss_cache;
    +       size_needed = probe_size;

Because

1. TCP associates a timestamp request with the last byte from send()
2. tcp_mtu_probe builds a probe from the oldest unacked bytes and
3. sized_needed exceeds probe_size

So the last byte in the current send call always > probe_size.

So this bug is rare and not easy to trigger deterministically from a
script without the above hack. But it may happen with more outstanding
unacked data.

Signed-off-by: Willem de Bruijn <willemb@google.com>